Move sql-session stuff to opengever.core.model
