+++ opengever/core/testing.py
@@ -5,9 +5,9 @@ from ftw.builder.testing import set_builder_session_factory
 from ftw.testing import ComponentRegistryLayer
 from ftw.testing.quickinstaller import snapshots
 from opengever.core import model
+from opengever.core.model import create_session
 from opengever.meeting.interfaces import IMeetingSettings
 from opengever.ogds.base.setup import create_sql_tables
-from opengever.ogds.base.utils import create_session
 from opengever.ogds.models import BASE
 from plone.app.testing import applyProfile
 from plone.app.testing import FunctionalTesting
