+++ opengever/globalindex/upgrades/to2602.py
@@ -1,5 +1,5 @@
 from ftw.upgrade import UpgradeStep
-from opengever.ogds.base.utils import create_session
+from opengever.core.model import create_session
 from opengever.ogds.models.utils import alter_column_length
