+++ opengever/ogds/base/hooks.py
@@ -1,7 +1,7 @@
 from ftw.dictstorage.sql import DictStorageModel
+from opengever.core.model import create_session
 from opengever.ogds.base.setup import create_example_client
 from opengever.ogds.base.setup import create_sql_tables
-from opengever.ogds.base.utils import create_session
 from opengever.ogds.models.group import Group
 from opengever.ogds.models.group import groups_users
 from opengever.ogds.models.org_unit import OrgUnit
