+++ opengever/ogds/base/setup.py
@@ -1,8 +1,8 @@
 from ftw.dictstorage.sql import DictStorageModel
-from opengever.ogds.base.utils import create_session
+from opengever.core.model import create_session
 from opengever.ogds.models import BASE
-from opengever.ogds.models.org_unit import OrgUnit
 from opengever.ogds.models.group import Group
+from opengever.ogds.models.org_unit import OrgUnit
 from z3c.saconfig.interfaces import IScopedSession
 from zope.component import queryUtility
