+++ opengever/ogds/base/sort_helpers.py
@@ -1,7 +1,7 @@
 from opengever.contact import contact_service
+from opengever.core.model import create_session
 from opengever.ogds.base.actor import Actor
 from opengever.ogds.base.interfaces import ISyncStamp
-from opengever.ogds.base.utils import create_session
 from opengever.ogds.base.utils import ogds_service
 from opengever.ogds.models.user import User
 from plone.memoize import ram
