+++ opengever/testing/sql.py
@@ -1,7 +1,7 @@
+from opengever.core.model import create_session
 from opengever.ogds.base.interfaces import ISyncStamp
 from opengever.ogds.base.ou_selector import CURRENT_ORG_UNIT_KEY
 from opengever.ogds.base.sync.import_stamp import update_sync_stamp
-from opengever.ogds.base.utils import create_session
 from opengever.ogds.base.utils import get_ou_selector
 from opengever.ogds.models.group import Group
 from opengever.ogds.models.user import User
